+++ src/gtt_menu_commands.c
@@ -34,7 +34,7 @@
 #include "props-proj.h"
 #include "timer.h"
 #include "toolbar.h"
-#include "xml-gtt.h"
+#include "gtt_xml.h"
 
 
 void
